Move around postgres code
