+++ src/DomainDriven/Persistance/Postgres.hs
@@ -1,0 +1,17 @@
+module DomainDriven.Persistance.Postgres
+    ( module X
+    ) where
+
+import           DomainDriven.Persistance.Postgres.Internal   as X
+                                                ( PostgresEvent(..)
+                                                , postgresWriteModel
+                                                , postgresWriteModelNoMigration
+                                                )
+import           DomainDriven.Persistance.Postgres.Types      as X
+                                                ( ChunkSize
+                                                , EventMigration
+                                                , EventTable(..)
+                                                , EventTableBaseName
+                                                , EventTableName
+                                                , PreviousEventTableName
+                                                )
